+++ envdata/views/__init__.py
@@ -1,0 +1,2 @@
+from .companies_views import *
+from .energy_views import *